backend/pyproject.toml
[project]
name = "course-companion-fte"
version = "1.0.0"
description = "Course Companion FTE — AI Agent Development course tutor backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.110.0",
    "uvicorn[standard]==0.27.1",
    "sqlalchemy==2.0.28",
    "alembic==1.13.1",
    "asyncpg==0.29.0",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "boto3==1.34.0",
    "python-multipart==0.0.9",
    "pydantic-settings==2.2.1",
    "httpx==0.27.0",
    "supabase>=2.28.3",
    "email-validator>=2.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.uv]
dev-dependencies = [
    "aiosqlite>=0.22.1",
    "httpx>=0.27.0",
    "pytest==8.1.0",
    "pytest-asyncio==0.23.5",
    "pytest-cov==5.0.0",
]
